[svn-r7611] Purpose:     Code cleanup
Description:
    Cleaned up various compiler warnings

Platforms tested:
    FreeBSD 4.9 (sleipnir)
    Linux 2.4 (verbena) w/FORTRAN
    too minor for h5committest

+++ fortran/src/H5Af.c
@@ -571,7 +571,7 @@ nh5aget_name_c(hid_t_f *attr_id, size_t_f *bufsize, _fcd buf)
       * Allocate buffer to hold name of an attribute
       */
      c_bufsize = *bufsize;
-     c_buf = (char *)HDmalloc((int)c_bufsize +1);
+     c_buf = (char *)HDmalloc(c_bufsize +1);
      if (c_buf == NULL) return ret_value; 
  
      /*
